requirements.txt.txt
pandas
numpy
scikit-learn
tensorflow-cpu
streamlit
joblib